requirements.txt
boto3[crt]
python-dotenv